feat: bundle ai credits
